Changed to a normaly .NET 4.5 assembly.
Trying to prove the problem with the build on MyGet is that it isn't
supporting portable libraries. If this works then will need to work with
the MyGet guys to see if they can introduce support for portable
libraries or figure out how to do a build some other way.